requirements.txt
boto3
psycopg2-binary
pandas
requests
dbt-core
dbt-postgres
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
great-expectations
prometheus-client
sqlalchemy